refactor(server): clean up the core module.
