+++ python/examples/simple_server.py
@@ -1,0 +1,17 @@
+"""A simple example of extending the ARFlow server."""
+
+import arflow
+
+
+class CustomService(arflow.ARFlowService):
+    def on_frame_received(self, frame: arflow.DataFrameRequest):
+        """Called when a frame is received."""
+        print("Frame received!")
+
+
+def main():
+    arflow.create_server(CustomService, port=8500, path_to_save="./")
+
+
+if __name__ == "__main__":
+    main()
